improve map popup
the popup used to open where the user clicked, which when zoomed out, used to be
far away. this made the google maps link not useful.
this commit changes the popup to open at the nearest point on the clicked
feature, so that the popup point is actually close to the desired spot